+++ projects/forms/src/public-api.ts
@@ -10,6 +10,7 @@ export * from './lib/show-form-control/show-form-control';
 export * from './lib/show-form-control/show-form-control.component';
 export * from './lib/forms-typed';
 export * from './lib/forms-util';
+export * from './lib/control-value-accessor-connector';
 
 @NgModule({
     imports: [CommonModule],
