+++ src/app/person-contact/person-contact.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Self, Optional } from '@angular/core';
 import { typedFormGroup, typedFormControl } from 'forms';
 import { PersonContact } from './person-contact.model';
-import { NgControl, ControlValueAccessor } from '@angular/forms';
-import { ControlValueAccessorConnector } from '../shared/control-value-accessor-connector';
+import { NgControl } from '@angular/forms';
+import { ControlValueAccessorConnector } from 'forms';
 
 @Component({
   selector: 'fty-person-contact',
